Membuat logika pemanggilan MyOrders dari DB kedalam JTable
