+++ TugasBesarPBO/src/Controller/CreateTable.java
@@ -10,8 +10,8 @@ package Controller;
 import java.sql.*;
 //import java.sql.DriverManager;
 //import java.sql.SQLException;
-import java.util.logging.Level;
-import java.util.logging.Logger;
+//import java.util.logging.Level;
+//import java.util.logging.Logger;
 
 /**
  *
