+++ TugasBesarPBO/src/View/MenuMyOrders.java
@@ -80,6 +80,7 @@ public class MenuMyOrders implements ActionListener{
 
     @Override
     public void actionPerformed(ActionEvent e) {
-        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
+        frame.setVisible(false);
+        new MenuMember();
     }
 }
